requirements.txt
flask
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
flask-sqlalchemy
flask-login
aiohttp
flask-jsonpify
markdown
markupsafe
bleach
itsdangerous
nltk
rake-nltk
google-images-search
google-search-results
azure-ai-textanalytics
